requirements.txt
hydra-core>=1.3
pytest>=8
PyYAML>=6.0
setuptools>=59
